backend/colabTeam_web/pyproject.toml
[tool.poetry]
name = "colabteam-web"
version = "0.0.1"
description = "Colab Team web based on Django framework."
authors = ["Your Name <you@example.com>"]
license = "New BSD"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
markdown = "^3.5"
django-cors-headers = "^4.3.0"
pillow = "^10.1.0"
drf-standardized-errors = "^0.12.6"
six = "^1.16.0"
cryptography = "^41.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
